fix(extra_fields): Fix issue with substream partition router picking extra fields (#727)
